Disable SIGPIPE on Swift socket transport
